REFACTORED Pose replaced Vector3, some floats replaced by doubles
Utils\MathHelper2.cs: double counterparts of some methods implemented

WaiterStupid:

McLocalization\Pose.cs: class represents pose of the robot (2d position vector and bearing angle)
McLocalization\LikelihoodFieldMeasurementModel.cs, McLrfLocalizator.cs, OdometryMotionModel.cs, PoseHistogram.cs, RangefinderProperties.cs: moved to Pose and doubles


git-svn-id: svn://brumba.ru:46104/brumba@142 cf9289f8-4b28-4085-b22d-9969776bd1f6

+++ src/Brumba/WaiterStupid/McLocalization/RangefinderProperties.cs
@@ -6,10 +6,10 @@ namespace Brumba.WaiterStupid.McLocalization
 {
     public struct RangefinderProperties
     {
-        public float AngularResolution { get; set; }
-        public float AngularRange { get; set; }
-        public float MaxRange { get; set; }
-        public float ZeroBeamAngleInRobot { get; set; }
+        public double AngularResolution { get; set; }
+        public double AngularRange { get; set; }
+        public double MaxRange { get; set; }
+        public double ZeroBeamAngleInRobot { get; set; }
 
         public Vector2 BeamToVectorInRobotTransformation(float zi, int i)
         {
@@ -22,7 +22,7 @@ namespace Brumba.WaiterStupid.McLocalization
             Contract.Requires(AngularResolution > 0);
             Contract.Ensures(Contract.Result<Vector2>().Length().AlmostEqualInDecimalPlaces(zi, 5));
 
-            return Vector2.Transform(new Vector2(zi, 0), Matrix.CreateRotationZ(ZeroBeamAngleInRobot + i * AngularResolution));
+            return Vector2.Transform(new Vector2(zi, 0), Matrix.CreateRotationZ((float)ZeroBeamAngleInRobot + i * (float)AngularResolution));
         }
     }
 }
